cleaning up schemas to better reflect that we are interfacing with the open stack identity service and not rackspace auth v2.0
