+++ project-set/components/client-auth/src/main/resources/META-INF/schema/config/client-auth-n-configuration.xsd
@@ -4,7 +4,7 @@
            targetNamespace="http://docs.rackspacecloud.com/power-api/client-auth/v1.0"
            xmlns:auth="http://docs.rackspacecloud.com/power-api/client-auth/v1.0"
            xmlns:auth-1.1="http://docs.rackspacecloud.com/power-api/client-auth/rs-auth-1.1/v1.0"
-           xmlns:auth-2.0="http://docs.rackspacecloud.com/power-api/client-auth/rs-auth-2.0/v2.0"
+           xmlns:os-ids-auth="http://docs.openrepose.org/power-api/client-auth/os-ids-auth/v1.0"
            xmlns:auth-http-basic="http://docs.rackspacecloud.com/power-api/client-auth/http-basic/v1.0"
            xmlns:html="http://www.w3.org/1999/xhtml"
            xmlns:xs="http://www.w3.org/2001/XMLSchema">
@@ -12,8 +12,8 @@
     <!-- Rackspace Cloud Auth v1.1 -->
     <xs:import namespace="http://docs.rackspacecloud.com/power-api/client-auth/rs-auth-1.1/v1.0" schemaLocation="./rackspace-auth-v1.1/rackspace-auth-v1.1.xsd" />
 
-    <!-- Rackspace Cloud Auth v2.0 -->
-    <xs:import namespace="http://docs.rackspacecloud.com/power-api/client-auth/rs-auth-2.0/v2.0" schemaLocation="./rackspace-auth-v2.0/rackspace-auth-v2.0.xsd" />
+    <!-- OpenStack Identity Service Auth -->
+    <xs:import namespace="http://docs.openrepose.org/power-api/client-auth/os-ids-auth/v1.0" schemaLocation="./openstack-ids-auth/openstack-ids-auth.xsd" />
     
     <!-- HTTP Basic Authentication -->
     <xs:import namespace="http://docs.rackspacecloud.com/power-api/client-auth/http-basic/v1.0" schemaLocation="./http-basic/http-basic.xsd" />    
@@ -31,10 +31,9 @@
         <xs:sequence>
             <xs:choice>
                 <xs:element ref="auth-1.1:rackspace-auth" />
-                <xs:element ref="auth-2.0:rackspace-auth-v2.0" />
+                <xs:element ref="os-ids-auth:openstack-auth" />
                 <xs:element ref="auth-http-basic:http-basic-auth" />
             </xs:choice>
         </xs:sequence>
-
     </xs:complexType>
 </xs:schema>
