+++ project-set/components/client-auth/src/main/resources/META-INF/schema/config/openstack-ids-auth/bindings.xjb
@@ -3,9 +3,9 @@
 <bindings xmlns="http://java.sun.com/xml/ns/jaxb" version="2.0"
         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
         xsi:schemaLocation="http://java.sun.com/xml/ns/jaxb http://java.sun.com/xml/ns/jaxb/bindingschema_2_0.xsd"
-        schemaLocation="rackspace-auth-v2.0.xsd">
+        schemaLocation="openstack-ids-auth.xsd">
 
     <schemaBindings>
-        <package name="com.rackspace.papi.components.clientauth.rackspace.config" />
+        <package name="com.rackspace.papi.components.clientauth.openstack.config" />
     </schemaBindings>
 </bindings>
