+++ project-set/external/service-clients/openstack-ids/pom.xml
@@ -1,19 +1,20 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">         
+    
+    <modelVersion>4.0.0</modelVersion>
+    
     <parent>
         <artifactId>service-client-support</artifactId>
         <groupId>com.rackspace.papi.external.clients</groupId>
         <version>1.0.3-SNAPSHOT</version>
     </parent>
-    <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.rackspace.cloud.services.clients</groupId>
-    <artifactId>rackspace-auth-v2.0</artifactId>
-    <version>1.0.3-SNAPSHOT</version>
-
-    <name>Power API Service Clients - Rackspace Auth v2.0</name>
+    <artifactId>openstack-ids</artifactId>
+    
+    <name>Power API Service Clients - OpenStack Identity Service</name>
 
     <packaging>jar</packaging>
 
@@ -24,8 +25,8 @@
         </dependency>
 
         <dependency>
-          <groupId>com.rackspace.papi.commons</groupId>
-          <artifactId>utilities</artifactId>
+            <groupId>com.rackspace.papi.commons</groupId>
+            <artifactId>utilities</artifactId>
         </dependency>
 
         <dependency>
@@ -44,15 +45,7 @@
             <plugin>
                 <groupId>org.codehaus.mojo</groupId>
                 <artifactId>xml-maven-plugin</artifactId>
-                <executions>
-                    <execution>
-                        <goals>
-                            <goal>transform</goal>
-                        </goals>
-                        <phase>generate-sources</phase>
-                    </execution>
-                </executions>
-
+               
                 <configuration>
                     <transformationSets>
                         <transformationSet>
@@ -60,7 +53,6 @@
                             <stylesheet>src/main/resources/META-INF/schema/xsl/s1.0p.xsl</stylesheet>
                         </transformationSet>
                     </transformationSets>                    
-
                 </configuration>
             </plugin>
 
@@ -69,12 +61,20 @@
                 <artifactId>maven-jaxb-plugin</artifactId>
 
                 <configuration>
-                    <schemaDirectory>${basedir}/target/generated-resources/xml/xslt</schemaDirectory>                
+                    <schemaDirectory>${basedir}/target/generated-resources/xml/xslt</schemaDirectory>           
+                    
+                    <includeSchemas>
+                        <includeSchema>**/*.xsd</includeSchema>
+                    </includeSchemas>
+                    
+                    <includeBindings>
+                        <includeBinding>**/*.xjb</includeBinding>
+                    </includeBindings>
+                    
                     <strict>true</strict>
                     <verbose>false</verbose>
                 </configuration>
             </plugin>
         </plugins>
     </build>
-
 </project>
