+++ project-set/external/service-clients/openstack-ids/src/main/java/com/rackspace/auth/openstack/ids/AuthenticationCache.java
@@ -1,4 +1,4 @@
-package com.rackspace.auth.v2_0;
+package com.rackspace.auth.openstack.ids;
 
 import net.sf.ehcache.CacheManager;
 import net.sf.ehcache.Element;
