+++ project-set/external/service-clients/openstack-ids/src/main/java/com/rackspace/auth/openstack/ids/CachableTokenInfo.java
@@ -1,4 +1,4 @@
-package com.rackspace.auth.v2_0;
+package com.rackspace.auth.openstack.ids;
 
 import org.openstack.docs.identity.api.v2.AuthenticateResponse;
 import org.openstack.docs.identity.api.v2.Role;
