+++ project-set/external/service-clients/openstack-ids/src/main/java/com/rackspace/auth/openstack/ids/ServiceClientResponse.java
@@ -1,4 +1,4 @@
-package com.rackspace.auth.v2_0;
+package com.rackspace.auth.openstack.ids;
 
 import java.io.InputStream;
