+++ project-set/external/service-clients/openstack-ids/src/main/resources/META-INF/schema/auth-2.0-service/atom/xml.xsd
