+++ project-set/external/service-clients/openstack-ids/src/test/java/com/rackspace/auth/v2_0/GenericServiceClientTest.java
