+++ project-set/external/service-clients/pom.xml
@@ -18,7 +18,7 @@
 
     <modules>
         <module>rackspace-auth-v1.1</module>
-        <module>rackspace-auth-v2.0</module>
+        <module>openstack-ids</module>
     </modules>  
 
     <dependencyManagement>
