+++ project-set/pom.xml
@@ -39,6 +39,21 @@
                 <pluginManagement>
                     <plugins>
                         <plugin>
+                            <groupId>org.codehaus.mojo</groupId>
+                            <artifactId>xml-maven-plugin</artifactId>               
+                            <version>1.0</version>
+                            
+                            <executions>
+                                <execution>
+                                    <goals>
+                                        <goal>transform</goal>
+                                    </goals>
+                                    <phase>generate-sources</phase>
+                                </execution>
+                            </executions>
+                        </plugin>         
+                        
+                        <plugin>
                             <groupId>com.sun.tools.xjc.maven2</groupId>
                             <artifactId>maven-jaxb-plugin</artifactId>
                             <version>1.1.1</version>
